Move ui impl back into lib.rs.
This was done to match the other modules. Also fixed line-feed issue
(normally don't work on windows).
